feat(StatisticsAPI) : 통계 api 1번~6번 개발
-이번달 통계 요약
-일주일간 비용 통계
-일주일간 충전량 통계
-완속 급속 비율(충전시간이 20초 이내면 급속)
-정상,고장 충전기 비율
-일주일간 충전소 전기요금 변동 통계
-api 테스트를 위한 데이터 삽입

+++ src/main/java/com/example/quaterback/api/feature/statistics/dto/response/StatisticsData.java
@@ -26,6 +26,6 @@ public class StatisticsData {
     public static class ChartData {
         private String id;     // pie chart 용
         private String label;
-        private int value;
+        private double value;
     }
 }
